+++ R/summaryfactorlist.R
@@ -13,13 +13,14 @@
 #' regression models.
 #'
 #' @param .data Dataframe.
-#' @param dependent Character vector of length 1:  name of dependent variable
-#'   (2 to 5 factor levels).
+#' @param dependent Character vector of length 1:  name of dependent variable (2
+#'   to 5 factor levels).
 #' @param explanatory Character vector of any length: name(s) of explanatory
 #'   variables.
 #' @param cont Summary for continuous variables: "mean" (standard deviation) or
 #'   "median" (interquartile range).
-#' @param cont_cut Numeric: number of unique values in continuous variable at which to consider it a factor.
+#' @param cont_cut Numeric: number of unique values in continuous variable at
+#'   which to consider it a factor.
 #' @param p Logical: Include statistical test (see
 #'   \code{\link[Hmisc]{summary.formula}}).
 #' @param na_include Logical: include missing data in summary (\code{NA}).
@@ -32,10 +33,11 @@
 #' @param na_to_missing Logical: convert \code{NA} to 'Missing' when
 #'   \code{na_include=TRUE}.
 #' @param add_dependent_label Add the name of the dependent label to the top
-#'   left of table
-#' @param dependent_label_prefix Add text before dependent label
-#' @param dependent_label_suffix Add text after dependent label
-#' @param ... Pass other arguments
+#'   left of table.
+#' @param dependent_label_prefix Add text before dependent label.
+#' @param dependent_label_suffix Add text after dependent label.
+#' @param ... Pass other arguments to \code{\link[Hmisc]{summary.formula}}),
+#'   e.g. \code{catTest = catTestfisher}.
 #' @return Returns a \code{factorlist} dataframe.
 #'
 #' @family finalfit wrappers
@@ -207,11 +209,11 @@ summary_factorlist_groups <- function(.data, dependent, explanatory,  cont = "me
 																			p = FALSE, na_include = FALSE,
 																			column = FALSE, total_col = FALSE, orderbytotal = FALSE, fit_id = FALSE,
 																			na_to_missing = TRUE, add_dependent_label = FALSE,
-																			dependent_label_prefix = "Dependent: ", dependent_label_suffix = ""){
+																			dependent_label_prefix = "Dependent: ", dependent_label_suffix = "", ...){
 	
 	s <- summary_formula(as.formula(paste(dependent, "~", paste(explanatory, collapse = "+"))), data = .data,
 											 method = "reverse", overall = TRUE,
-											 test = TRUE, na.include = na_include, continuous = cont_cut)
+											 test = TRUE, na.include = na_include, continuous = cont_cut, ...)
 	df.out = plyr::ldply(1:length(s$stats), function(index) {
 		x = s$stats[[index]]
 		is_continuous = s$type[index] == 2