+++ tests/testthat/test_tests.R
@@ -1,0 +1,6 @@
+context("Statistical tests")
+library(finalfit)
+test_that("catTestfisher works", {
+	expect_is(summary_factorlist(colon_s, "mort_5yr", "age.factor", catTest = catTestfisher), "data.frame")
+})
+